+++ packages/at_commons/CHANGELOG.md
@@ -1,5 +1,8 @@
-## 3.0.58
+## 4.0.0
 - fix: Add "put" operation to OTP verb to store semi-permanent pass codes
+## 3.0.58
+- fix: Deprecate encryptedDefaultEncryptedPrivateKey in EnrollParams and introduce encryptedDefaultEncryptedPrivateKey for readability
+- fix: Replace encryptedDefaultEncryptedPrivateKey with encryptedDefaultEncryptionPrivateKey in EnrollVerbBuilder
 ## 3.0.57
 - feat: Introduced TTL(Time to Live) for OTP verb to configure OTP expiry
 ## 3.0.56